+++ DefaultClicker/Assets/Scripts/ShopItems/ShopItemScorePerSec.cs
@@ -31,7 +31,7 @@ namespace SekiburaGames.DefaultClicker.UI
 
         }
 
-        private void OnPriceUpdate(float nextBGPrice)
+        private void OnPriceUpdate(double nextBGPrice)
         {
             if (_priceText != null)
                 _priceText.text = nextBGPrice.ToString() + "$";
